refactor: change icons dir
Attempt on resolving tsc error in CI
