+++ docs/about-us/contact-us.mdx
@@ -5,7 +5,7 @@ title: Contact Us
 
 import Link from '@docusaurus/Link';
 import Logo from '@site/src/components/uis/Logo';
-import { IcGithub, IcYoutube, IcTwitter, IcCopy, ImgLocation } from '@site/src/utils/icons.ts';
+import { IcGithub, IcYoutube, IcTwitter, IcCopy, ImgLocation } from '@site/src/components/icons.ts';
 import { DOOBOOLAB_EMAIL } from '@site/src/utils/constants';
 import { copyEmailAsync } from '@site/src/utils/functions';
