+++ docs/about-us/introduction.mdx
@@ -2,7 +2,7 @@
 sidebar_position: 1
 ---
 
-import { IcLogoIntroLoop } from '../../src/utils/icons';
+import { IcLogoIntroLoop } from '@site/src/components/icons';
 
 <IcLogoIntroLoop />
