+++ i18n/ko/docusaurus-plugin-content-docs/current/about-us/introduction.mdx
@@ -4,7 +4,7 @@ sidebar_label: 소개
 title: 소개
 ---
 
-import { IcLogoIntroLoop } from '@site/src/utils/icons';
+import { IcLogoIntroLoop } from '@site/src/components/icons';
 
 <IcLogoIntroLoop />
