+++ src/components/icons.ts
